+++ GSTBillingApp/Models/OwnerListViewModel.cs
@@ -8,7 +8,7 @@ namespace GSTBillingApp.Models
         public int Id { get; set; }
         public string OwnerName { get; set; }
         public string GSTNo { get; set; }
-        public int ContactNo { get; set; }
+        public long ContactNo { get; set; }
         public string OwnerAddress { get; set; }
     }
 }
